+++ android/src/main/java/com/klarna/inapp/sdk/PaymentViewWrapper.java
@@ -1,9 +1,7 @@
 package com.klarna.inapp.sdk;
 
 import android.util.AttributeSet;
-import android.util.Log;
 import android.view.View;
-import android.webkit.ValueCallback;
 import android.webkit.WebView;
 import android.widget.LinearLayout;
 
@@ -15,93 +13,88 @@ import com.klarna.mobile.sdk.api.payments.KlarnaPaymentView;
 /***
  * Wraps the KlarnaPaymentView so we can see when a requestLayout() has been triggered.
  */
-public class PaymentViewWrapper extends LinearLayout {
-
+public class PaymentViewWrapper extends LinearLayout implements HeightListener.HeightListenerCallback {
     private float displayDensity = 1;
     public KlarnaPaymentView paymentView;
+    private boolean loadCalled = false;
+    private HeightListener heightListener;
 
     public PaymentViewWrapper(ReactApplicationContext context, AttributeSet attrs) {
         super(context, attrs);
-
         // Get density for resizing.
-        displayDensity = getResources().getDisplayMetrics().density;
-
+        displayDensity = context.getResources().getDisplayMetrics().density;
         // Add KlarnaPaymentView
         LinearLayout.LayoutParams webViewParams = new LinearLayout.LayoutParams(LinearLayout.LayoutParams.MATCH_PARENT, LayoutParams.MATCH_PARENT);
         paymentView = new KlarnaPaymentView(getReactAppContext().getCurrentActivity(), attrs); // Insure we use activity and not application context for dialogs.
         addView(paymentView, webViewParams);
+        heightListener = new HeightListener(getPaymentViewWebView(), this);
+    }
+
+    public void load(String sessionData) {
+        paymentView.load(sessionData);
+        loadCalled = true;
+        heightListener.injectListener(getPaymentViewWebView());
+    }
+
+    @Override
+    public void requestLayout() {
+        super.requestLayout();
+        if (isReady() && loadCalled) {
+            heightListener.fetchHeight(getPaymentViewWebView());
+        }
     }
 
     /**
      * The web view inside the KlarnaPaymentView will trigger requestLayout(), but there is no way
      * to get the KlarnaPaymentView's correct height.
-     *
+     * <p>
      * Attempts include:
      * - Using Android's measure() and co. methods
      * - Getting the WebView's getContentHeight()
-     * - Using a variety of layout listners.
-     *
+     * - Using a variety of layout listeners.
+     * <p>
      * So instead we're:
      * 1. Evaluating some JS (yes, making things even more fragile).
      * 2. Triggering a size change via the UIManagerModule. This will effectively apply a fixed size
-     *    via width and height style attributes on the React side of things.
-     *
+     * via width and height style attributes on the React side of things.
+     * <p>
      * Note: We can apply style by setting props with uimm.updateView() to just set a height (and
-     *       no width), but it doesn't update immediately or with the right height.
-     *
+     * no width), but it doesn't update immediately or with the right height.
+     * <p>
      * The width will be whatever the parent component's width is.
      */
-    @Override
-    public void requestLayout() {
-        super.requestLayout();
-
-        if (!isReady()) {
+    private void setHeight(String height) {
+        if (height == null || height.equals("null") || height.equals("undefined")) {
             return;
         }
-
-        WebView innerWebView = getPaymentViewWebView();
-
-        if (innerWebView == null) {
-            return;
+        try {
+            final int width = getParentViewWidth();
+            final float contentHeight = Float.parseFloat(height);
+            final float scaledHeight = contentHeight * displayDensity;
+            getReactAppContext().runOnNativeModulesQueueThread(new GuardedRunnable(getReactAppContext()) {
+                @Override
+                public void runGuarded() {
+                    UIManagerModule uimm = getReactAppContext().getNativeModule(UIManagerModule.class);
+                    uimm.updateNodeSize(getId(), width, (int) scaledHeight);
+                }
+            });
+        } catch (Throwable t) {
         }
-
-        final int viewWidth = getParentViewWidth();
-
-        // Note: window.innerHeight, document.body.getClientBoundingRect(), etc... don't return the
-        // right value.
-        innerWebView.evaluateJavascript("document.body.scrollHeight", new ValueCallback<String>() {
-            @Override
-            public void onReceiveValue(String value) {
-                if(value == null || value.equals("null") || value.equals("undefined")){
-                    return;
-                }
-                try {
-                    final float contentHeight = Float.parseFloat(value);
-                    final float scaledHeight = contentHeight * displayDensity;
-                    getReactAppContext().runOnNativeModulesQueueThread(new GuardedRunnable(getReactAppContext()) {
-                        @Override
-                        public void runGuarded() {
-                            UIManagerModule uimm = getReactAppContext().getNativeModule(UIManagerModule.class);
-                            uimm.updateNodeSize(getId(), viewWidth, (int) scaledHeight);
-                        }
-                    });
-                } catch (Throwable t){
-                    return;
-                }
-            }
-        });
     }
 
     /**
      * Returns true if the payment view is part of the hierarchy (the first requestLayout() is
      * triggered before) and the whole view is part of the RN hierarchy.
+     *
      * @return true if ready
      */
     private boolean isReady() {
         return paymentView != null && getId() != -1;
     }
 
-    /** Returns the parent view's width. */
+    /**
+     * Returns the parent view's width.
+     */
     private int getParentViewWidth() {
         View parentReactView = (View) getParent();
         if (parentReactView == null || !(parentReactView instanceof View)) {
@@ -110,7 +103,9 @@ public class PaymentViewWrapper extends LinearLayout {
         return parentReactView.getWidth();
     }
 
-    /** Attempts to retrieve the web view from within the KlarnaPaymentView. This is really fragile. */
+    /**
+     * Attempts to retrieve the web view from within the KlarnaPaymentView. This is really fragile.
+     */
     private WebView getPaymentViewWebView() {
         for (int i = 0; i < paymentView.getChildCount(); i++) {
             View child = paymentView.getChildAt(i);
@@ -122,8 +117,15 @@ public class PaymentViewWrapper extends LinearLayout {
         return null;
     }
 
-    /** Returns the app context the wrapper was initialized with. */
+    /**
+     * Returns the app context the wrapper was initialized with.
+     */
     private ReactApplicationContext getReactAppContext() {
         return (ReactApplicationContext) getContext();
+    }
+
+    @Override
+    public void onNewHeight(String value) {
+        setHeight(value);
     }
 }