+++ integration-tests/src/test/kotlin/com/testapp/model/SessionInfoResponse.kt
@@ -1,0 +1,14 @@
+package com.testapp.model
+
+internal data class SessionInfoResponse(
+        val session: Session
+)
+
+internal data class Session(
+        var client_token: String?,
+        var payment_method_categories: List<PaymentMethodCategory>
+)
+
+internal data class PaymentMethodCategory(
+        var identifier: String
+)