+++ CHANGELOG.md
@@ -1,17 +1,41 @@
 # Changelog
 All notable changes to React Native Klarna In-App SDK will be documented in this file.
 
+### [2.0.14] - 2020-10-30
+- Update iOS SDK version to the latest (2.0.28)
+
+### [2.0.13] - 2020-10-21
+- Update Android SDK version to the latest (2.0.28)
+
+### [2.0.12] - 2020-10-21
+- Change Android SDK version to 2.0.26
+
+### [2.0.11] - 2020-10-08
+- Update iOS SDK version to the latest (2.0.24)
+
+### [2.0.10] - 2020-09-10
+- Update iOS SDK version to the latest (2.0.21)
+
+### [2.0.9] - 2020-07-14
+- Updated web view height calculation on Android.
+
+### [2.0.8] - 2020-06-29
+- Fixed a resizing issue on Android when the web view's height was not properly updated.
+
+### [2.0.7] - 2020-06-12
+- Updated the iOS SDK version to the latest (2.0.17)
+
 ### [2.0.6] - 2020-06-08
-- Updated the parameters of reauthorize and finalize methods. 
+- Updated the parameters of reauthorize and finalize methods.
 
 ### [2.0.5] - 2020-05-27
-- Fixed a resizing issue on Android when the web view's scroll height value was returned null. 
+- Fixed a resizing issue on Android when the web view's scroll height value was returned null.
 
 ### [2.0.4] - 2020-05-06
 - Updated the iOS SDK version to the latest (2.0.13)
 
 ### [2.0.3] - 2020-04-27
-- Fixed a rendering issue while loading the payment view in Android SDK 
+- Fixed a rendering issue while loading the payment view in Android SDK
 
 ### [2.0.2] - 2020-04-20
 - Updated the iOS SDK version to the latest (2.0.12)